Funcionalidad Food

+++ GestionPae/src/main/java/com/example/GestionPae/model/User.java
@@ -17,7 +17,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idUser;
+    private Long idUser;
 
     private String name;
